+++ src/main/java/nextstep/app/config/FilterConfig.java
@@ -2,14 +2,14 @@ package nextstep.app.config;
 
 import java.util.List;
 import nextstep.app.MemberService;
-import nextstep.security.AuthenticationManager;
-import nextstep.security.BasicAuthenticationFilter;
-import nextstep.security.DaoAuthenticationProvider;
-import nextstep.security.DefaultSecurityFilterChain;
-import nextstep.security.FilterChainProxy;
-import nextstep.security.LoginFormAuthenticationFilter;
-import nextstep.security.ProviderManager;
-import nextstep.security.SecurityFilterChain;
+import nextstep.security.authentication.AuthenticationManager;
+import nextstep.security.filter.BasicAuthenticationFilter;
+import nextstep.security.authentication.DaoAuthenticationProvider;
+import nextstep.security.filter.DefaultSecurityFilterChain;
+import nextstep.security.filter.FilterChainProxy;
+import nextstep.security.filter.LoginFormAuthenticationFilter;
+import nextstep.security.authentication.ProviderManager;
+import nextstep.security.filter.SecurityFilterChain;
 import nextstep.security.context.HttpSessionSecurityContextRepository;
 import nextstep.security.context.SecurityContextHolderFilter;
 import org.springframework.context.annotation.Bean;
