+++ src/main/java/nextstep/security/UserDetailsService.java
@@ -1,7 +1,5 @@
 package nextstep.security;
 
-import nextstep.security.domain.UserDetails;
-
 public interface UserDetailsService {
 
     UserDetails findByUsername(String username);
