+++ src/main/java/nextstep/security/context/SecurityContextImpl.java
@@ -1,6 +1,6 @@
 package nextstep.security.context;
 
-import nextstep.security.domain.Authentication;
+import nextstep.security.authentication.Authentication;
 
 public class SecurityContextImpl implements SecurityContext {
