+++ src/main/java/nextstep/security/filter/BasicAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.filter;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
@@ -7,21 +7,24 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.app.util.Base64Convertor;
+import nextstep.security.authentication.AuthenticationManager;
 import nextstep.security.context.SecurityContext;
 import nextstep.security.context.SecurityContextHolder;
-import nextstep.security.domain.Authentication;
-import nextstep.security.domain.UsernamePasswordAuthenticationToken;
+import nextstep.security.authentication.Authentication;
 
 import java.io.IOException;
-import java.util.Map;
+import nextstep.security.authentication.AuthenticationException;
+import nextstep.security.authentication.UsernamePasswordAuthenticationToken;
 
-public class LoginFormAuthenticationFilter implements Filter {
+public class BasicAuthenticationFilter implements Filter {
 
     public static final String AUTHORIZATION = "Authorization";
+    private static final String BASIC_AUTH_PREFIX = "Basic ";
 
     private final AuthenticationManager authenticationManager;
 
-    public LoginFormAuthenticationFilter(AuthenticationManager authenticationManager) {
+    public BasicAuthenticationFilter(AuthenticationManager authenticationManager) {
         this.authenticationManager = authenticationManager;
     }
 
@@ -34,23 +37,20 @@ public class LoginFormAuthenticationFilter implements Filter {
 
         final String authorization = ((HttpServletRequest) request).getHeader(AUTHORIZATION);
 
-        if (!isDefaultAuthentication(authorization)) {
+        if (!isBasicAuthentication(authorization)) {
             chain.doFilter(request, response);
             return;
         }
 
-        final Map<String, String[]> parameterMap = request.getParameterMap();
-        Authentication authentication = null;
+        Authentication authentication;
 
         try {
-            authentication = authenticateUser(parameterMap);
-        } catch (RuntimeException e) {
-            SecurityContextHolder.clearContext();
-            ((HttpServletResponse) response).setStatus(401);
-            return;
-        }
+            authentication = authenticateUser(authorization);
 
-        if (!authentication.isAuthenticated()) {
+            if (!authentication.isAuthenticated()) {
+               throw new AuthenticationException();
+            }
+        } catch (IllegalArgumentException | AuthenticationException e) {
             SecurityContextHolder.clearContext();
             ((HttpServletResponse) response).setStatus(401);
             return;
@@ -71,27 +71,37 @@ public class LoginFormAuthenticationFilter implements Filter {
     }
 
     /**
-     * 기본, Login Form 형식의 인증 정보인지 확인합니다.
+     * Basic 형식의 인증 정보인지 확인합니다.
      */
-    private boolean isDefaultAuthentication(final String authorization) {
-        return authorization == null;
+    private boolean isBasicAuthentication(final String authorization) {
+        return authorization != null && authorization.startsWith(BASIC_AUTH_PREFIX);
     }
 
     /**
      * 인증이 유효한 Authentication 객체를 반환합니다.
      */
-    private Authentication authenticateUser(final Map<String, String[]> parameterMap) {
-        String username;
-        String password;
-
-        try {
-            username = parameterMap.get("username")[0];
-            password = parameterMap.get("password")[0];
-        } catch (RuntimeException e) {
-            throw new IllegalArgumentException(e);
-        }
+    private Authentication authenticateUser(final String authorization) throws IllegalArgumentException {
+        final String[] usernameAndPassword = getUsernameAndPassword(authorization);
+        final String username = usernameAndPassword[0];
+        final String password = usernameAndPassword[1];
 
         return authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(username, password));
+    }
+
+    /**
+     * authorization 정보 기반으로, username과 password를 반환합니다.
+     */
+    private String[] getUsernameAndPassword(final String authorization) {
+        final String credentials = authorization.split(" ")[1];
+        final String decodedString = Base64Convertor.decode(credentials);
+
+        final String[] usernameAndPassword = decodedString.split(":");
+
+        if (usernameAndPassword.length != 2) {
+            throw new IllegalArgumentException();
+        }
+
+        return decodedString.split(":");
     }
 
     /**
