+++ src/test/java/nextstep/app/FormLoginTest.java
@@ -3,7 +3,7 @@ package nextstep.app;
 import jakarta.servlet.http.HttpSession;
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import nextstep.security.domain.Authentication;
+import nextstep.security.authentication.Authentication;
 import nextstep.utils.TestUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
